now travis will pass, apearantly templates don't play nice with definitions in a cpp file so all is defined and declared in the hpp file
