+++ code/main.cpp
@@ -1,7 +1,6 @@
 #include <hwlib.hpp>
 
 #include <hardware_usart.hpp>
-
 int main(void) {
   // kill the watchdog
     WDT->WDT_MR = WDT_MR_WDDIS;
